requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
loguru==0.7.2
celery[redis]
redis
uagents>=0.13.0
uagents-ai>=0.4.0
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
